Add Javadoc to QuestionTest

+++ aon-occam/src/test/java/com/esferalia/aon/occam/test/marketing/question/QuestionAONTest.java
@@ -17,60 +17,63 @@ import org.junit.Test;
 import com.esferalia.aon.occam.api.model.Question;
 import com.esferalia.aon.occam.api.model.QuestionParams;
 import com.esferalia.aon.occam.impl.jooq.RegistryImpl;
+import com.esferalia.aon.occam.impl.jooq.validation.QuestionValidation;
 import com.esferalia.aon.occam.test.AbstractOccamTest;
 import com.esferalia.aon.occam.test.Asserts;
 import com.esferalia.aon.occam.test.faker.AonFaker;
 import com.esferalia.aon.occam.api.AON;
 
-
+/**
+ * Tests the question's methods of the class AON
+ */
 public class QuestionAONTest extends AbstractOccamTest {
 
+	/**
+	 * Test create and delete
+	 */
 	@Test
 	public void crudeTest() {
-		// create
+		// create and update
 		Question question = AonFaker.getQuestion(ctx);
-		AON.saveQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question);
-		Asserts.assertEqualsQuestion(AON.getQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question.getId()),question);
-		
-		// delete
-		AON.deleteQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question.getId());
-		assertNull(AON.getQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question.getId()));
+		if (!QuestionValidation.checkAlias(ctx, question)) {
+			AON.saveQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question);
+			Asserts.assertEqualsQuestion(AON.getQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question.getId()),question);
+			
+			// delete
+			AON.deleteQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question.getId());
+			assertNull(AON.getQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question.getId()));
+		}
 	}
 	
+	/**
+	 * Test that it throws and exception if we get a list without the params that we set
+	 */
 	@Test
 	public void checkQuestionListTest() {
 		Question question1 = AonFaker.getQuestion(ctx);
 		Question question2 = AonFaker.getQuestion(ctx);
 		
-		question1.setAlias("alias");
-		question2.setAlias("alias");
+		question1.setActive(true);
+		question2.setActive(false);
 		
+		AON.saveQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question1);
+		AON.saveQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question2);
+
 		QuestionParams params = new QuestionParams()
-				.setAlias("alias")
 				.setDomain(ctx.getDomainId())
 				.setDomainName(ctx.getDomainName())
 				.setUser(ctx.getUser())
+				.setActive((byte) 1)
 				;
 		
-				/*
-				 * // if the alias does not exists if
-				 * (!AON.checkQuestionAlias(ctx.getDomainName(), ctx.getDomainId(),
-				 * ctx.getUser(), "alias")) { assertTrue(AON.getQuestionList(params).isEmpty());
-				 * 
-				 * AON.saveQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(),
-				 * question1); AON.saveQuestion(ctx.getDomainName(), ctx.getDomainId(),
-				 * ctx.getUser(), question2);
-				 * 
-				 * assertFalse(AON.getQuestionList(params).isEmpty());
-				 * 
-				 * AON.deleteQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(),
-				 * question1.getId()); AON.deleteQuestion(ctx.getDomainName(),
-				 * ctx.getDomainId(), ctx.getUser(), question2.getId());
-				 * 
-				 * assertTrue(AON.getQuestionList(params).isEmpty());
-				 * 
-				 * } else { assertFalse(AON.getQuestionList(params).isEmpty()); }
-				 */
+		List<Question> actual = AON.getQuestionList(params);
+		
+		for (Question question : actual) {
+			assertTrue(question.isActive());
+		}
+		
+		AON.deleteQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question1.getId());
+		AON.deleteQuestion(ctx.getDomainName(), ctx.getDomainId(), ctx.getUser(), question2.getId());
 	}
 
 }
